projects/basic-fileupload/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
PyYAML==6.0.1
Werkzeug==2.3.7
python-socketio==5.9.0
eventlet==0.33.3